% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_top_N}
\alias{select_top_N}
\title{Select top N rows from a data.table/ data.frame}
\usage{
select_top_N(input, N, decreasing = TRUE)
}
\arguments{
\item{input}{The input data.table/ data.frame}

\item{N}{The number of rows to select}

\item{decreasing}{Whether to select the top N rows or the bottom N rows}
}
\value{
The the selected indices of the top N rows
}
\description{
Select top N rows from a data.table/ data.frame
}
\examples{
dt.input <- data.frame(a = c(1, 2, 3), b = c(4, 5, 6))
select_top_N(dt.input, 2)
select_top_N(dt.input, 2, decreasing = FALSE)
}
