% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{selectSite}
\alias{selectSite}
\title{Select Site for the given file
p-Value, t-Stats, abds_diffs}
\usage{
selectSite(file_name)
}
\arguments{
\item{file_name}{The file name of the data}
}
\description{
Select Site for the given file
p-Value, t-Stats, abds_diffs
}
\examples{
file_name = tempfile()
data(demo_devEpt_1k)
write.table(demo_devEpt_1k[1:10, ], file_name, row.names=F,  sep = '\t', quote = FALSE)
uu <- selectSite(file_name)
}
